+++ android/guava/src/com/google/common/reflect/Invokable.java
@@ -21,8 +21,8 @@ import com.google.common.collect.ImmutableList;
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
 import java.lang.annotation.Annotation;
 import java.lang.reflect.AccessibleObject;
+import java.lang.reflect.AnnotatedElement;
 import java.lang.reflect.Constructor;
-import java.lang.reflect.GenericDeclaration;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Member;
 import java.lang.reflect.Method;
@@ -49,18 +49,29 @@ import org.checkerframework.checker.nullness.qual.Nullable;
  * assertEquals(new TypeToken<List<String>>() {}, invokable.getOwnerType());
  * }</pre>
  *
+ * <p><b>Note:</b> earlier versions of this class inherited from {@link
+ * java.lang.reflect.AccessibleObject AccessibleObject} and {@link
+ * java.lang.reflect.GenericDeclaration GenericDeclaration}. Since version 31.0 that is no longer
+ * the case. However, most methods from those types are present with the same signature in this
+ * class.
+ *
  * @param <T> the type that owns this method or constructor.
  * @param <R> the return type of (or supertype thereof) the method or the declaring type of the
  *     constructor.
  * @author Ben Yu
- * @since 14.0
+ * @since 14.0 (no longer implements {@link AccessibleObject} or {@code GenericDeclaration} since
+ *     31.0)
  */
 @Beta
 @ElementTypesAreNonnullByDefault
-public abstract class Invokable<T, R> extends Element implements GenericDeclaration {
+public abstract class Invokable<T, R> implements AnnotatedElement, Member {
+  private final AccessibleObject accessibleObject;
+  private final Member member;
 
   <M extends AccessibleObject & Member> Invokable(M member) {
-    super(member);
+    checkNotNull(member);
+    this.accessibleObject = member;
+    this.member = member;
   }
 
   /** Returns {@link Invokable} of {@code method}. */
@@ -71,6 +82,151 @@ public abstract class Invokable<T, R> extends Element implements GenericDeclarat
   /** Returns {@link Invokable} of {@code constructor}. */
   public static <T> Invokable<T, T> from(Constructor<T> constructor) {
     return new ConstructorInvokable<T>(constructor);
+  }
+
+  @Override
+  public final boolean isAnnotationPresent(Class<? extends Annotation> annotationClass) {
+    return accessibleObject.isAnnotationPresent(annotationClass);
+  }
+
+  @Override
+  @CheckForNull
+  public final <A extends Annotation> A getAnnotation(Class<A> annotationClass) {
+    return accessibleObject.getAnnotation(annotationClass);
+  }
+
+  @Override
+  public final Annotation[] getAnnotations() {
+    return accessibleObject.getAnnotations();
+  }
+
+  @Override
+  public final Annotation[] getDeclaredAnnotations() {
+    return accessibleObject.getDeclaredAnnotations();
+  }
+
+  // We ought to be able to implement GenericDeclaration instead its parent AnnotatedElement.
+  // That would give us this method declaration. But for some reason, implementing
+  // GenericDeclaration leads to weird errors in Android tests:
+  // IncompatibleClassChangeError: interface not implemented
+  /** See {@link java.lang.reflect.GenericDeclaration#getTypeParameters()}. */
+  public abstract TypeVariable<?>[] getTypeParameters();
+
+  /** See {@link java.lang.reflect.AccessibleObject#setAccessible(boolean)}. */
+  public final void setAccessible(boolean flag) {
+    accessibleObject.setAccessible(flag);
+  }
+
+  /** See {@link java.lang.reflect.AccessibleObject#trySetAccessible()}. */
+  public final boolean trySetAccessible() {
+    // We can't call accessibleObject.trySetAccessible since that was added in Java 9 and this code
+    // should work on Java 8. So we emulate it this way.
+    try {
+      accessibleObject.setAccessible(true);
+      return true;
+    } catch (RuntimeException e) {
+      return false;
+    }
+  }
+
+  /** See {@link java.lang.reflect.AccessibleObject#isAccessible()}. */
+  public final boolean isAccessible() {
+    return accessibleObject.isAccessible();
+  }
+
+  @Override
+  public final String getName() {
+    return member.getName();
+  }
+
+  @Override
+  public final int getModifiers() {
+    return member.getModifiers();
+  }
+
+  @Override
+  public final boolean isSynthetic() {
+    return member.isSynthetic();
+  }
+
+  /** Returns true if the element is public. */
+  public final boolean isPublic() {
+    return Modifier.isPublic(getModifiers());
+  }
+
+  /** Returns true if the element is protected. */
+  public final boolean isProtected() {
+    return Modifier.isProtected(getModifiers());
+  }
+
+  /** Returns true if the element is package-private. */
+  public final boolean isPackagePrivate() {
+    return !isPrivate() && !isPublic() && !isProtected();
+  }
+
+  /** Returns true if the element is private. */
+  public final boolean isPrivate() {
+    return Modifier.isPrivate(getModifiers());
+  }
+
+  /** Returns true if the element is static. */
+  public final boolean isStatic() {
+    return Modifier.isStatic(getModifiers());
+  }
+
+  /**
+   * Returns {@code true} if this method is final, per {@code Modifier.isFinal(getModifiers())}.
+   *
+   * <p>Note that a method may still be effectively "final", or non-overridable when it has no
+   * {@code final} keyword. For example, it could be private, or it could be declared by a final
+   * class. To tell whether a method is overridable, use {@link Invokable#isOverridable}.
+   */
+  public final boolean isFinal() {
+    return Modifier.isFinal(getModifiers());
+  }
+
+  /** Returns true if the method is abstract. */
+  public final boolean isAbstract() {
+    return Modifier.isAbstract(getModifiers());
+  }
+
+  /** Returns true if the element is native. */
+  public final boolean isNative() {
+    return Modifier.isNative(getModifiers());
+  }
+
+  /** Returns true if the method is synchronized. */
+  public final boolean isSynchronized() {
+    return Modifier.isSynchronized(getModifiers());
+  }
+
+  /** Returns true if the field is volatile. */
+  final boolean isVolatile() {
+    return Modifier.isVolatile(getModifiers());
+  }
+
+  /** Returns true if the field is transient. */
+  final boolean isTransient() {
+    return Modifier.isTransient(getModifiers());
+  }
+
+  @Override
+  public boolean equals(@CheckForNull Object obj) {
+    if (obj instanceof Invokable) {
+      Invokable<?, ?> that = (Invokable<?, ?>) obj;
+      return getOwnerType().equals(that.getOwnerType()) && member.equals(that.member);
+    }
+    return false;
+  }
+
+  @Override
+  public int hashCode() {
+    return member.hashCode();
+  }
+
+  @Override
+  public String toString() {
+    return member.toString();
   }
 
   /**
@@ -165,13 +321,12 @@ public abstract class Invokable<T, R> extends Element implements GenericDeclarat
   @SuppressWarnings("unchecked") // The declaring class is T's raw class, or one of its supertypes.
   @Override
   public final Class<? super T> getDeclaringClass() {
-    return (Class<? super T>) super.getDeclaringClass();
+    return (Class<? super T>) member.getDeclaringClass();
   }
 
   /** Returns the type of {@code T}. */
   // Overridden in TypeToken#method() and TypeToken#constructor()
   @SuppressWarnings("unchecked") // The declaring class is T.
-  @Override
   public TypeToken<T> getOwnerType() {
     return (TypeToken<T>) TypeToken.of(getDeclaringClass());
   }
